requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.3.5
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2
SQLAlchemy==2.0.23
requests==2.31.0
python-dateutil==2.8.2
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
gunicorn==21.2.0
twilio==8.10.0
python-multipart==0.0.6 